backend/requirements.txt
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
apache-superset==4.1.1
apispec==6.3.0
apsw==3.46.0.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
billiard==4.2.1
blinker==1.9.0
Bottleneck==1.4.2
Brotli==1.1.0
build==1.2.1
cachelib==0.9.0
cachetools==5.5.0
camelot-py==0.11.0
cattrs==24.1.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-option-group==0.5.6
click-plugins==1.1.1
click-repl==0.3.0
cmdstanpy==1.1.0
colorama==0.4.6
contourpy==1.0.7
coverage==7.6.8
cron-descriptor==1.4.5
croniter==5.0.1
cryptography==42.0.8
cycler==0.12.1
db-dtypes==1.3.1
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
docker==7.0.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
Faker==33.1.0
filelock==3.12.2
filetype==1.2.0
Flask==2.3.3
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-Compress==1.17
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Migrate==3.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==2.5.1
flask-talisman==1.1.0
Flask-Testing==0.8.1
Flask-WTF==1.2.2
fonttools==4.55.0
freezegun==1.5.1
func_timeout==4.3.5
future==1.0.0
geographiclib==2.0
geopy==2.4.1
gevent==24.2.1
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.155.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.19.1
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.0.3
grpcio==1.68.0
grpcio-status==1.60.1
gunicorn==23.0.0
h11==0.14.0
hashids==1.3.1
holidays==0.25
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
humanize==4.11.0
identify==2.5.36
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.17.3
jsonschema-spec==0.1.6
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kombu==5.4.2
korean-lunar-calendar==0.3.1
langchain-core==0.3.25
langchain-google-genai==2.0.7
langsmith==0.2.3
lazy-object-proxy==1.10.0
limits==3.13.0
llvmlite==0.43.0
lxml==5.3.0
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.9.0
mdurl==0.1.2
msgpack==1.0.8
msgspec==0.18.6
mysqlclient==2.2.6
nh3==0.2.19
nodeenv==1.8.0
numba==0.60.0
numexpr==2.10.2
numpy==1.23.5
oauthlib==3.2.2
odfpy==1.4.1
openapi-schema-validator==0.4.4
openapi-spec-validator==0.5.6
openpyxl==3.1.5
ordered-set==4.1.0
orjson==3.10.12
outcome==1.3.0.post0
packaging==24.2
pandas==2.0.3
pandas-gbq==0.19.1
parameterized==0.9.0
paramiko==3.5.0
parsedatetime==2.6
pathable==0.4.3
pdfminer.six==20231228
pdfplumber==0.11.4
pgsanity==0.2.9
pillow==10.3.0
pip-compile-multi==2.6.3
pip-tools==7.4.1
platformdirs==3.8.1
playwright==1.42.0
pluggy==1.5.0
ply==3.11
polyline==2.0.2
pre_commit==4.0.1
prison==0.2.1
progress==1.6
prompt_toolkit==3.0.48
prophet==1.1.5
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.0
psycopg2-binary==2.9.6
pyarrow==14.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydata-google-auth==1.9.0
pydruid==0.6.9
pyee==11.0.1
pyfakefs==5.3.5
Pygments==2.18.0
PyHive==0.7.0
pyinstrument==4.4.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
pypdf==5.1.0
pypdfium2==4.30.0
pyproject_hooks==1.2.0
pyrsistent==0.20.0
PySocks==1.7.1
pytesseract==0.3.13
pytest==7.4.4
pytest-cov==6.0.0
pytest-mock==3.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-geohash==0.8.5
python-ldap==3.4.4
pytz==2024.2
pyxlsb==1.0.10
PyYAML==6.0.2
redis==4.6.0
referencing==0.35.1
requests==2.32.2
requests-cache==1.2.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.8.0
selenium==3.141.0
shillelagh==1.2.18
shortid==0.1.2
simplejson==3.19.3
six==1.16.0
slack_sdk==3.33.4
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.54
sqlalchemy-bigquery==1.12.0
SQLAlchemy-Utils==0.38.3
sqlglot==25.24.5
sqloxide==0.1.51
sqlparse==0.5.2
sshtunnel==0.4.0
statsd==4.0.1
tabulate==0.8.10
tenacity==9.0.0
tomli==2.1.0
toposort==1.10
tqdm==4.67.1
trino==0.330.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.18
vine==5.1.0
virtualenv==20.23.1
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
WTForms==3.2.1
WTForms-JSON==0.3.5
xlrd==2.0.1
XlsxWriter==3.0.9
zipp==3.21.0
zope.event==5.0
zope.interface==5.4.0
zstandard==0.23.0
